refactor: Создать компоненты Header, Footer
