+++ src/components/favorites-screen/favorites-screen.jsx
@@ -1,59 +1,20 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
-import {connect} from 'react-redux';
+import Header from '../layout/header';
+import Footer from '../layout/footer';
 import FavoritesList from './favorites-list';
-import {user as userType} from '../../types';
-import {AppRoute} from '../../const';
 
-const FavoritesScreen = ({user}) => {
+const FavoritesScreen = () => {
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={AppRoute.ROOT}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={AppRoute.FAVORITES}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">
-                      { user.email }
-                    </span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <FavoritesList />
         </div>
       </main>
-      <footer className="footer container">
-        <a className="footer__logo-link" href="main.html">
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width={64} height={33} />
-        </a>
-      </footer>
+      <Footer />
     </div>
   );
 };
 
-FavoritesScreen.propTypes = {
-  user: userType,
-};
-
-const mapStateToProps = (state) => ({
-  user: state.user,
-});
-
-export {FavoritesScreen};
-export default connect(mapStateToProps, null)(FavoritesScreen);
+export default FavoritesScreen;
