+++ src/components/layout/header.jsx
@@ -1,0 +1,56 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import {Link} from 'react-router-dom';
+import {connect} from 'react-redux';
+import {user as userType} from '../../types';
+import {AuthorizationStatus, AppRoute} from '../../const';
+
+const Header = ({authorizationStatus, user}) => {
+  return (
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <div className="header__left">
+            <Link className="header__logo-link header__logo-link--active" to={AppRoute.ROOT}>
+              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
+            </Link>
+          </div>
+          <nav className="header__nav">
+            <ul className="header__nav-list">
+              <li className="header__nav-item user">
+                {
+                  authorizationStatus === AuthorizationStatus.AUTH ?
+                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.FAVORITES}>
+                      <div className="header__avatar-wrapper user__avatar-wrapper">
+                      </div>
+                      <span className="header__user-name user__name">
+                        { user.email }
+                      </span>
+                    </Link> :
+                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.LOGIN}>
+                      <div className="header__avatar-wrapper user__avatar-wrapper">
+                      </div>
+                      <span className="header__login">Sign in</span>
+                    </Link>
+                }
+              </li>
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+};
+
+Header.propTypes = {
+  authorizationStatus: PropTypes.string.isRequired,
+  user: userType,
+};
+
+const mapStateToProps = (state) => ({
+  authorizationStatus: state.authorizationStatus,
+  user: state.user,
+});
+
+export {Header};
+export default connect(mapStateToProps, null)(Header);
