+++ src/components/main-screen/main-screen.jsx
@@ -1,18 +1,17 @@
 import React, {useEffect} from 'react';
 import PropTypes from 'prop-types';
-import {Link} from 'react-router-dom';
 import {connect} from 'react-redux';
 // import {ActionCreator} from '../../store/action';
+import Header from '../layout/header';
 import CitiesPlaces from '../cities/cities-places';
 import NoPlaces from '../cities/no-places';
 import Map from '../map/map';
 import LocationsList from '../locations/locations-list';
-import {offers as offersType, user as userType} from '../../types';
+import {offers as offersType} from '../../types';
 import LoadingScreen from '../loading-screen/loading-screen';
 import {fetchOffers} from '../../store/api-actions';
-import {AuthorizationStatus, AppRoute} from '../../const';
 
-const MainScreen = ({authorizationStatus, user, currentOffers, isOffersLoaded, onLoadOffers}) => {
+const MainScreen = ({currentOffers, isOffersLoaded, onLoadOffers}) => {
 
   useEffect(() => {
     if (!isOffersLoaded) {
@@ -36,39 +35,7 @@ const MainScreen = ({authorizationStatus, user, currentOffers, isOffersLoaded, o
 
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link header__logo-link--active" to={AppRoute.ROOT}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  {
-                    authorizationStatus === AuthorizationStatus.AUTH ?
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.FAVORITES}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        <span className="header__user-name user__name">
-                          { user.email }
-                        </span>
-                      </Link> :
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.LOGIN}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        <span className="header__login">Sign in</span>
-                      </Link>
-                  }
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className={mainPageClass}>
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
@@ -96,16 +63,12 @@ const MainScreen = ({authorizationStatus, user, currentOffers, isOffersLoaded, o
 };
 
 MainScreen.propTypes = {
-  authorizationStatus: PropTypes.string.isRequired,
-  user: userType,
   currentOffers: offersType,
   isOffersLoaded: PropTypes.bool.isRequired,
   onLoadOffers: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
-  user: state.user,
   currentOffers: state.currentOffers,
   isOffersLoaded: state.isOffersLoaded,
 });
