+++ src/components/offer-screen/offer-screen.jsx
@@ -1,21 +1,20 @@
 import React, {useEffect} from 'react';
-import {Link} from 'react-router-dom';
 import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
 import {calcRatingProgress} from '../../utils/offers.js';
+import Header from '../layout/header';
 import ReviewsBlock from '../reviews/reviews-block';
 import User from '../user/user';
 import NearOffersList from './near-offers-list';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
-import {offer as offerType, user as userType} from '../../types';
+import {offer as offerType} from '../../types';
 import OfferMap from './offer-map';
 import LoadingScreen from '../loading-screen/loading-screen';
-import {AuthorizationStatus, AppRoute} from '../../const';
 import {fetchOffer} from '../../store/api-actions';
 
 const MAX_IMAGES = 6;
 
-const OfferScreen = ({authorizationStatus, user, currentOffer, isOfferLoaded, onLoadOffer, match: {params: {id}}}) => {
+const OfferScreen = ({currentOffer, isOfferLoaded, onLoadOffer, match: {params: {id}}}) => {
   const idInt = parseInt(id, 10);
 
   useEffect(() => {
@@ -49,39 +48,7 @@ const OfferScreen = ({authorizationStatus, user, currentOffer, isOfferLoaded, on
 
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={AppRoute.ROOT}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  {
-                    authorizationStatus === AuthorizationStatus.AUTH ?
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.FAVORITES}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        <span className="header__user-name user__name">
-                          { user.email }
-                        </span>
-                      </Link> :
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.LOGIN}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        <span className="header__login">Sign in</span>
-                      </Link>
-                  }
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--property">
         <section className="property">
           <div className="property__gallery-container container">
@@ -156,14 +123,6 @@ const OfferScreen = ({authorizationStatus, user, currentOffer, isOfferLoaded, on
               <div className="property__host">
                 <h2 className="property__host-title">Meet the host</h2>
                 <User user={ host } className={`property__host-user`} />
-                {/* <div className="property__host-user user">
-                  <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
-                    <img className="property__avatar user__avatar" src={host.avatarUrl} width={74} height={74} alt="Host avatar" />
-                  </div>
-                  <span className="property__user-name">
-                    { host.name }
-                  </span>
-                </div> */}
                 <div className="property__description">
                   <p className="property__text">
                     { description }
@@ -188,16 +147,12 @@ const OfferScreen = ({authorizationStatus, user, currentOffer, isOfferLoaded, on
 
 OfferScreen.propTypes = {
   match: PropTypes.object.isRequired,
-  authorizationStatus: PropTypes.string.isRequired,
-  user: userType,
   currentOffer: offerType,
   isOfferLoaded: PropTypes.bool.isRequired,
   onLoadOffer: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
-  user: state.user,
   currentOffer: state.currentOffer,
   isOfferLoaded: state.isOfferLoaded,
 });
